Allow failure of Newspeak on linux 64
This architecture and flavour are failing for a long time
Such failures make the CI status useless
and obstruct integration of unrelated features

It is vital to have a green CI, and these are provisional measures to restore a sane status.
This should not prevent Newspeak teams to investigate the problem further and hopefully resolve it.