Cog VM source as per VMMaker.oscog-eem.2561
Fix a bug in followForwardedLiteralsIn:; whether the methodObject is young or
not is important, and must be recorded.

Fix Float nan comparison again...
It was broken in  VMMaker.oscog-nice.2557
The comment of broken method tells what it should do (and what it did),
but no more what it does

Fix awfull VM crash when testing
[48 = $0] bench.

genJumpImmediate: generates not one instruction but two (Compare + Jump).
So we cannot use it as a jumpTarget (or we skip the Compare instruction!)