Minor patch for making unix SocketPlugin 64bits friendly
IPV4 address are uint32_t, not u_long
This is not the same on 64 bits.
This has a consequence on alignment of pointer aliasing (u_long might require 8 bytes alignment).
This remove a runtime UndefinedBehavior sanitizer message (UBsan)
when compiled with CC=clang ./mvm and CFLAGS -fsanitize=undefined.
So far, the consequence are void, but let's not insult the future...

Note that I used autotools HAVE_STDINT_H for backward compatibility
but we could either drop this compatibility like already done in SSL or Camera plugin...
C99 is already 20years behind!

Also avoid converting a pointer (PSP) to (unsigned long) just for printing...
We have %p now for that purpose (like already used elsewhere in the same file).