CogVM source as per VMMaker.oscog-eem.2568
CoInterpreter:
Fix a slip in printing contexts.

Cogit:
Eliminate a bogus assert in compileFullBlockMethodFrameBuild: (and explain why).

Stop generating forwarder following code for constants and self in methods in
followed because self can be a forwarder in blocks without inst var accesses
(anyway self == something in block is not super common).