CogVM source as per VMMaker.oscog-eem.2570
Cog:
Refactor to make CFramePointer and CStackPointer private variables of
CoInterpreter.  Hence change the signature of isCFramePointerInUse, which now
takes pointers to the two variables instead of referencing them directly.  The
result is that CFramePointer and CStackPointer are accessed via VarBaseReg on
relevant platforms.  This nearly halves the size of the generated
trampolines/enilopmarts on x86_64.

Fix a reg arg order overwrite problem with the ceDirectedSuperSend?Args
trampolines on ARM32.

Slang: Fix a bug in inferring the type of addressOf:[put:].