Don't declare functions as volatile
This is absolutely pointless in C.
The compiler could put the call outside a loop whether declared volatile or not, if it can figure that the function has no side effect.
If the function reads volatile variables, then the compiler can't optimize and is forced to re-read in each loop.