Nuke recode entirely from unix X11
`recode` is used to translate the `charCode`... charCode is keyboard `event at: 3`.
From which encoding to which encoding? Err... charCode should be a keycode (keysym on X11).
So it should not be UTF8 or whatever. It's a keycode not a character code (code point).
Beside, charCode will use a single sqInt event slot, and a single byte is expected by image side.

Currently with my `fr_FR.UTF-8` locale, the VM is trying to translate a multi-byte encoded Character into 1-byte `charCode`.
This does not make sense.

Once upon a time, we used to encode the character into macRoman encoding, and used only the charCode at image side.
Now we should always use the ucs4 `event at: 6`, and we should STOP translating the charCode.
The charCode should be renamed keycode, and should just be that, a platform independent key code.
(including left shift, right shift, etc...).
key codes are useful for making lower level handlers at image side, handling key up and key down events.