Use translateCode primarily for translating symbolic X11 keysym to OSVM keycode
For this purpose, enhance `translateCode` to handle classical Tab/Esc/Return/BS/Del keys (otherwise, we get 16rFFxx keycodes)
Note: currently, we do not differentiate Return from KeyPad Enter keys...

Move `mapDelBs` hack in `translateCode`.
Note: IMO we should nuke this heresy once for all...
It only make sense if we build the VM input layer above a deficient programming layer. We shall better not do that!
The added complexity is minor, but how many such useless "features" do we carry, for what added value exactly?

Note: if `translateCode` does not handle the keysym (it `return -1`) then we try to convert the key into a character using `XLookupString`.
If it converts to one (or more) byte characters (`buf` may be encoded UTF8 for example), then simply use the `*symbolic` keysym.
Else return -1, meaning that we do not recognize the key.

I did not change that last part, one change at a time, but it might be a better idea to just pass the `*symbolic` verbatim to the image in all cases.