Convert OSX point units into scrolling units
Set the limit to 20 scrolling units to restrain the flow of events delivered to image
That's about 7 points (1/10 inch), not exactly the smoothest.
But if we want smoother scroll, we have to accelerate event handling at image side first.

The threshold might become a parameter, but for now YAGNI.

Note: the scrolling unit of 120 units per mouse wheel notch comes from moz.dev.
The fact that we convert a single notch to 3 lines is a Squeak convention
See https://github.com/OpenSmalltalk/opensmalltalk-vm/issues/41