Fix the signature of b3dDrwBufferFunction
We instantiate this function pointer with a function taking `sqInt` rather than `int`.
In 32 bits VM, it's OK, but on 64bits VM, sqInt is 64bits wide.
This prototype mismatch did prevent sign extension of negative int values.
Thus an offset of -1 in `yValue` becomes an offset of 2^32-1 in the callee,
which then cause a buffer overrun (and most often a SEGV)