CogVM source as per VMMaker.oscog-eem.2671
Spur:
Succumb to temptation and avoid moving the value through the FPU in 64-bit
floatAt:[put:].  Rename the primitives to fit the house style
(primitiveSpurFoo, not primitiveFooSpur).
Refactor smallFloatObjectOf: to avoid duplication.

Cogit:
Refactor simulation of flushICacheFrom:to: so that it actually runs on
processors which we can and do choose to implement this in machine code.  The
implementation is now in the backEnd and defers to the simulator if required.

Compute numTrampolines more accurately, deferring to the backEnd class for
backEnd specific trampolines such as ceCheckFeatures.

Add AddCqRR (useful for ARMv8 cache flushing).

Clean up the definition of the optional trampolines for ZeroCount, low-level
locks, processor feature checks and cache flushing.