CogVM source as per VMMaker.oscog-eem.2673
Placate Slang by adding a macro for the internal vs external tag on the end of
the plugin moduleName, so that we keep the changes in VMMaker.oscog-eem.2671 &
VMMaker.oscog-eem.2672 for hack #if declarations, and don't have to add yet
another hack to handle the old style (and more verbose) moduleName declaration.
And move declareModuleName: to VMPluginCodeGenerator where it belongs.

Alas this means that we regenerate all plugins, but just this once.

Fix the unofficial exampleSqNamedPrims.h files in some lowcode dirs.