CogVM source as per Name: VMMaker.oscog-eem.2675
Cogit: Correct the definition of genPrimitiveHighBit.  If there is no
implementation of LZCNT then it should return UnimplementedPrimitive, not
CompletedPrimitive.  Reverse the order of operands in ClzR:R:; this is a
left-to-right assembler, not an ATT syntax right-to-left one.
Implement OrCq:R:R: to save an instruction in genPrimitiveHighBit.
Move genHighBitIn:ofSmallIntegerOopWithSingleTagBit: into the backEnd where it
belongs.  Nuke the useless
genHighBitAlternativeIn:ofSmallIntegerOopWithSingleTagBit:.

ARMv8 linux/unix:
Make sure there are compilable definitions of getfp & getsp and that the back
trace handler can get hold of fp & sp.
Builds:
Add build directories for squeak.cog.spur on ARMv8.  Modify the mvm scripts to
install build.linux64ARMv8 builds under a name that doesn't conflict with
build.linux64x64.