Instrument the Squeak3D fill list operations with additional consistency checks
Removing a face which is absent from the fill list will break the fill list.
Indeed, fill list is a doubly linked list, and remove function is reconnecting prevFace and nextFace of removed face.
But if this face has already been removed, we reuse dangling prevFace/nextFace and break the linked list...

Same for adding: if the face was already on the list, we ignore the prevFace/nextFace links and thus break the fill list too.

Note that this change just instrument, not fix.