CogVM source as per VMMaker.oscog-eem.2719
Slang: type left shifts as either usqInt or sqInt, depending on type of
receiver, not #int.  These are Smalltalk semantics we're trying to mimic, not
C99 semantics we're constrained to follow.

Consequently also generate all plugin source (that changes), which implies
changes due to VMMaker.oscog-eem.2697, allow the CoInterpreter to specify the
clustering of variables by size.

Cogit: don't bother to rewrite the selector in a lnked super send; this causes
overwriting of the selector index in 64-bit implementations, and hence may
cause an assert failure.  So it's a waste of effort and provokes an error.