Finally compile with a more recent compiler than gcc-4.2 or llvm-gcc-4.2
Note: gcc-4.2 fails to compile the spur VM with an internal error.
I guess we ain't gonna get any support for fixing it ;)

With XCode 10.3 and XcodeLegacy https://github.com/devernay/xcodelegacy,
I could compile on macos 10.14 and run in both osx 10.6 and 10.14

It's not ideal to not be able to compile directly from the target OS, but that's the best we can get for supporting this legacy thing.
It's already nice enough to have latest spur VM running on (snow) leopard!