CogVm source as per VMMaker.oscog-eem.2722
Interpreter:
Fix compilation of primitiveHighBit on MSVC.  We take the address of a variable
by using (self addressOf: var), not "var address".

Change the signature of the core selector send breakpointing routines from
taking the receiver to taking a class tag. (Requires corresponding changes in
platforms/Cross/vm/dispdbg.h).
Add an accessor for breakLookupClassTag and add couldBeContext:.  Comment
fetchClassTagOf:. [This was all to track down a bug with ARMv8 do do with
register save/restore across the call in ceScheduleScavengeTrampoline that
caused contexts to appear where blocks were expected].

Fix printMethodFieldForPrintContext: to always print the method oop frst and
the cogMethod, if any, afterwards.

Fix asserts in frameCallerContext: and mapStackPages when a frame context may
be forwarded as part of scavenging.

Cogit:
rename the accessor for codeToDataDelta to getCodeToDataDelta to allow
codeToDataDelta to be defined as 0 in the non DUAL_MAPPED_CODE_ZONE regime.
Consequently find and fix a slip in NewspeakCogMethod class>>initialize.
Don't attempt to take the address of 0 in the DUAL_MAPPED_CODE_ZONE regime.
Fix genLoadCStackPointer(s) to use NativeSPReg

Fix followForwardedLiteralsIn:, mapObjectReferencesInMachineCodeForXXX, and
storeLiteral:atAnnotatedAddress:using: for the DUAL_MAPPED_CODE_ZONE regime.

Don't bother to rewrite the selector in a linked super send; this causes
overwriting of the selector index in 64-bit implementations, and hence may
cause an assert failure.  So it's a waste of effort and provokes an error.

Plugins:

Fix the BitBlt bug for convex shape fill.

The problem is that we are taking source one word too early (thus from previous
row).  This happens when taking hdir=-1 and it is related to preload (again!!).

Last thing: on 64bits, mask1 and mask2 have been declared int64...
But we only want 32 bit masks.
That is why I have protected with the bitAnd: AllOnes