CogVM source as per VMMaker.oscog-eem.2731
Cogit:
Refactor the compilation breakpoints so they don't take a selector length.
Leave it up to CoInterpeeter to derive the byte length of the selector using
numBytesOf:.  Hence no longer mark lengthOf: as <api>.

Use a writable method when setting cpicHasMNUCase: in
cogExtendPIC:CaseNMethod:tag:isMNUCase:.  ARMv8 is getting there...

Slang: do constant folding in left and right shifts.

Interpreters: rip out the check alloc filler support in the interests of
simplicity.  Though useful in theory, this faclity has never been used
successfully to find a bug in all the years it's been available (and we
can always put it back).
Rip out the debugPrimCallStackOffset support.  I can't remember what this
was supposed to do.  It is confusing.

Spur: Don't update become effect flags for identical oops that will later
be filtered out in the loops oiver the arrays.  If newFinalizartion is set,
then don't queue a WeakArray for finalization more than once (i.e. if it's
already in the queue there's no point adding it again).

Cogit: fix genExternalizePointersForPrimitiveCall &
genLoadCStackPointersForPrimCall for the SPReg ~= NativeSPReg regime.
Localise code to the primitive generator invocation block in compilePrimitive.