CogVM source as per VMMaker.oscog-eem.2736
CoInterpreter: No longer rely on setjmp/longjmp to get back into the interpreter from arbitrary locatons.  Instead, since CoInterpreter maintains the base of the C stack in CFramePointer & CStackPointer, it is straight-forward for us to simply call interpret after doing the switch to the C stack, avoiding issues such as stack unwind problems in longjmp.  And of course the implementation is simpler thn setjmp/longjmp and so faster."

Eliminate reenterInterpreter in the CoInterpreter (keeping it in the StackInterpreter; althoguh the success of the use in the CoInterpreter shows us that we should do the same, e.g. using assembler, in the StackInterpreter).

So now reentry into the interpreter from machine code does not use longjmp and we no longer use setjmp to establish the reentry-point.  We just use CFramePointer & CStackPointer.

Export the warning functions for Windows DLLs that want to use assertions.
Get rid of the preambleCCode implementation of invalidCompactClassError:

Cogit x64, fix a slip in computing the size of SignExtend32RR.

Fix bad bad copy/past error in generateSignedShiftRight:on:indent:

Generic:
Fix a slip in oopForPointer in the USE_INLINE_MEMORY_ACCESSORS regime when
sqMemoryBase is not defined or is non-zero; one cannot perform arithmetic on
void pointers so the argument must be cast to char.

Bochs simulators: eliminate the TLB; it is not used in our configuration.

FileAttributesPlugin generated as per FileAttributesPlugin.oscog-eem.54