CogVM source as per VMMaker.oscog-eem.2737
CoInterpreter: Improve on the new simpler invoke interpreter scheme by including
the relevant return address so that the C stack looks properly connected after
reentering interpret.  This depends on getReturnAddress being defined in C.
Both MSVC and GCC (and those that support its builtins such as clang and ICC)
provide builtins/intrinsics for accessing the return address, so we don't bother
to implement it in the Cogit.

The implementation of getReturnAddress is in sq.h, not sqPlatformSpecific.h to
reduce duplication.  But if the current implementations in terms of MSVC or GCC
instrinsics don't apply, one may override in sqPlatformSpecific.h.

Windows builds:
Make most of the windows build makefiles switch hit between the cygwin/mingw
and the MSVC variants depending on a variable set by MSVC command prompts.
Hence to build using MSVC simply run make in an MSVC prompt.