CogVM source as per VMMaker.oscog-eem.2751
Spur: Fix assert that checks for a valid function pointer on primitive failure
that was confiused by FFI calls.  Fix potential bug with primitive failure of
external and FFI calls, making sure the first literal of the method is followed.
The first literal is state used to cache an external call or define the
signature of an ffi call.  Make sure that FFI calls have a valid accessor depth.

Fix a common speeling rorre.

Slang: initGlobalStructure is never used; nuke its generator.