CogV source as per VMMaker.oscog-eem.2761 Author: eem Time: 24 June 2020, 11:15:52.846414 am UUID: 34b29eae-6069-4cb7-8a89-7365fb398dfb Ancestors: VMMaker.oscog-eem.2760
Cherry Picking from various recent commits, but avoiding the extremely desirable, but as yet unfinished, VMMaker.oscog-nice.2761 transformInAssignmentTo: changes.the "thisContext method includes: 42" crash.

MiscPrimitivePlugin: fix several uses of sizeOfSTArrayFromCPrimitive: that don''t check for potential failure if e.g. invoked on a CompiledMethod.  Returning from a primitive normally when the primtiive has failed leads to disaster since the stack gets cu back but shouldn't be.  For CompiledMethod isBytes: is true but isWordsOrBytes: is false.  sizeOfSTArrayFromCPrimitive: checks for isWordsOrBytes:.  The primtiives check that the argument is isBytes: but don't check if sizeOfSTArrayFromCPrimitive: fails.  More general fixes, such as fixing isBytes: to be false for CompiledMethod, or introducing isPureBytes: and using it, are not quick fixes.  hence this limited fix here.

Primitive infrastructure: consequently guard the various methodReturnXXX''s with an assert to check that a primitive has not failed.

Also make sure that the SpurNBitCoMemoryManagers do not follow any reference to a Cog method in the first field of a CompiledMethod.

Cosmetic changes for ThreadedFFIPlugins from VMMaker.oscog-nice.2762

Do not try to generate SHA256Plugin, it's obsolete and absent from latest cryptography packages. This from VMMaker.oscog-nice.2761.

Slang:
Fix TParseNode>>isSameAs: implementations to incluyde an identity check.  TReturnNode always answered false to this in the past.

Optimize inlineFunctionCall:in: to avoid a rewrite of the copied parse tree being inlined if the actuals match the formals.  Uses the improved bindVariablesIn:.

Use a setter for variable & expression in TAssignmentNode to ease breakpointing/debugging.

Scripts: Fix revertIfEssentiallyUnchanged to handle the threaded FFI plugins
now they include version info in the moduleName routine.