CogVM source as per VMMaker.oscog-eem.2773/ClosedVMMaker-eem.93
Cogit:
Fix the use of the undefined ceFlushDCache function when DUAL_WRITE_CODE_ZONE is not defined and flushDCacheFrom:to: is not a no-op.
Make sure any generated check features, and the generated run-time are cache flushed before use.
Move maybeGenerateCheckLZCNT later initializeCodeZoneFrom:upTo:; its result is not used until compiling primitives.