CogVM source as per VMMaker.oscog-eem.2792
General:
Eliminate translation time type error warnings for extendedStoreBytecodePop:
and fetchLong32:ofFloatObject:.

Simplify and correct the comments of some of the integer oop => value
conversion routines.  These routines can simply return values directly
instead of assigning through a variable.

Spur:
Looking at class indentityHash distributions in my current VMMaker image it is
clear that setting the classTableIndex to point at the start of the last used
page is not a good strategy, and leads to far too sparse a class table.  So
change the policy and set the classTableIndex to the first unused slot.

Eliminate unintentional duplication in
inLineRunLeakCheckerFor:excludeUnmarkedObjs:classIndicesShouldBeValid: