CogVM source as per VMMaker.oscog-eem.2798
Interpreters:
Clean up setjmp/longjmp.  Use only _setjmp:/_longjmp:_:.  To get Win64 working
it is easier to use a setjmp/longjmp pair that does not do stack unwinding.
It is more difficult to stitch teh stack properly so that Kernel32's checking
on stack unwinding does not raise an error.  So now we're using _setjmp/longjmp
everywhere and can fix things on WIN32 either to to invoke longjmpex or to use
a _setjmp/_longjmp replacement a la Win64.
[ci skip]