CogVM source as per FileAttributesPlugin.oscog-eem.57
Recind the rush of blood to the head.  Since the specialObjectsArray was only
changed in Squeak in January make primitiveFileMasks use WordArray only when
available (a WordArray is smaller and faster to access).  Also, thanks to Pierce
fix the regression in forgetting to export the first switch-hitting version.

Use common code for string return.