CogVM source as per VMMaker.oscog-eem.2831 Author: eem Time: 29 September 2020, 8:01:48.222992 pm UUID: f4fed277-6b62-4e79-9147-c5405589f2b1 Ancestors: VMMaker.oscog-eem.2830
Cogit: (IAAM) correctly implement the fix in VMMaker.oscog-eem.2824.  The
previous version of the fix used methodHeaderOf: to get what was intended
to be the pointer to the cog method, if newMethod was in fact cogged, but
methodHeaderOf: always answers the byetcoded method header.  What is needed
here is the actual pointer to the cog method, i.e. rawHeaderOf:.

CameraPlugin MacOS, add the AVCaptureSessionPreset352x288 size.
win32: put function names at the start of the line.