CameraPlugin source as per VMMaker.oscog-eem.2840
Add a primCameraUID primitive, to equal facilities in the Qwaq QWebcam plugin
(even though this appears inoperative on Windows).
Add module start-up and shutdown operations, useful for cleanup.
On Windows fix a storage leak with enumerating names.
On Linux implement the plumbing for CameraSetSemaphore and check the necessary
capability, but at least on 5.x kernels and earliuer there is no SIGIO notifi-
cation, and hence the asynchronous interface is currently unavailable.  We could
work around this by spawning a thread to do blocking reads, but I don't have the
time right now.
Make a function static in sqUnixSocket.c.
Correct the ioctl code (unused on current platforms) to use the TIOCSPGRP arg
to ioctl (SIOCSPGRP is socket-specific). This may be wrong, but if it is it will
require the aioEnable interface to be extended to allow clients to state whether
they are sockets or not.
Enable the CameraPlugin on build.linux64ARMv8.