CogVM/IA32API plugin source as per VMMaker.oscog-eem.2847
Fix a bad bug in IA32ABIPlugin>>#primReturnAsFromContextThrough, which must
always signal the semaphore whether it returns to C oir not, because if more
than one callback is waiting to retuern and one that fails to return doesn't
signal the semaphore to give anothe rwaiitng callback the chance to return
the system will deadlock.
Add savedMostRecentCallbackContext to VMCallbackContext so we can get the
entire list from the image to debug deadlocks.
Consign IA32ABIPlugin>>#primReturnFromContextThrough to history.