CogVM source as per VMMaker.oscog-eem.2849
Fix an awful regression introduced in VMMaker.oscog-eem.2837.  Have the
setFullScreen fag also set the flag that is stored in the image header,
hence having the image start up as it was saved.
And have the flags not recognised by a specific VM still get recorded in
imageHeaderFlags, so set parameters will get set in the image header.