IA32ABI & ThreadedFFIPlugins as per VMMaker.oscog-eem.2855
Add oopAt:[put:] support to Alien, needed for callbacks that exchange Smalltalk
objects.  We need to discuss the security/safety implications, but the basic
facility is essential for object-to-object interconnects such as JNI, and
indeed Qwaq's PyBridge Python interconnect.

Use a slightly more efficient isAlien: in IA32ABI and SqueakFFIPrims.